Rename benchmarks to benchmarks container, move child benchmarks into seperate component to allign with naming scheme
